requirements.txt
asgiref==3.3.1
attrs==20.3.0
autobahn==20.12.3
Automat==20.2.0
bleach==3.2.1
certifi==2020.12.5
cffi==1.14.4
channels==3.0.3
chardet
constantly==15.1.0
cryptography==3.3.1
daphne==3.0.1
defusedxml==0.6.0
Django==3.1.4
django-allauth==0.44.0
django-autoslug==1.9.8
django-bootstrap-datepicker-plus==3.0.5
django-crispy-forms==1.10.0
django-el-pagination==3.3.0
django-extensions==3.1.0
django-filter==2.4.0
django-location-field==2.1.0
django-markdownfield==0.8.0
django-post-office==3.5.3
easy-thumbnails==2.7.1
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
jsonfield==3.1.0
Markdown==3.3.3
mysqlclient==2.0.2
oauthlib==3.1.0
packaging==20.8
Pillow==8.0.1
pip-review==1.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
PyMySQL==0.10.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
python3-openid==3.2.0
pytz==2020.4
requests
requests-oauthlib==1.3.0
service-identity==18.1.0
shortuuid==1.0.1
six==1.15.0
sqlparse==0.4.1
Twisted==20.3.0
txaio==20.12.1
urllib3==1.26.2
webencodings==0.5.1
zope.interface==5.2.0
